% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_country_parameters.r
\name{gta_country_parameters}
\alias{gta_country_parameters}
\title{convert between different country parameters
Use this function to convert between different country attributes such as from name to un code or from un code to iso code.
Conversion is possible between all parameters recorded in the \code{gtalibrary::country.correspondence} table.}
\usage{
gta_country_parameters(
     c("Switzerland", "Norway", "Slovenia"),
     from = c(name, un_code, iso_code, iso.code.2letter, ...),
     to = c(name, un_code, iso_code, iso.code.2letter, ...),
     message = TRUE,
     keep_input = FALSE
)
}
\arguments{
\item{input}{a vector which contains the parameters to be converted. The type of parameters recorded in the vector
must be specified in the from argument}

\item{from}{Same parameters as in to}

\item{to}{un_code, iso.code.2letter, iso_code, name (+ all other colnames from the dataset gtalibrary::country.correspondence)}

\item{message}{Specify whether you want to receive a version about the success of conversion}

\item{keep_input}{if \code{TRUE}, the function returns a tibble where one column corresponds to the input vector and the second
column to the matching parameter specified in \code{to}}
}
\description{
convert between different country parameters
Use this function to convert between different country attributes such as from name to un code or from un code to iso code.
Conversion is possible between all parameters recorded in the \code{gtalibrary::country.correspondence} table.
}
