% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cpc code expand.r
\name{gta_cpc_code_expand}
\alias{gta_cpc_code_expand}
\title{Expands 2-digit level CPC codes into 3-digit level CPC codes.}
\usage{
gta_cpc_code_expand(
     codes,
     as_list = FALSE,
     message = TRUE
)
}
\arguments{
\item{codes}{Supply the 2-digit level CPC codes you want to expand.}

\item{as_list}{indicate whether the converted results are returned as a list or as a vector of unique results (default = FALSE)}

\item{message}{indicate whether conversion results are printed to the console or not (default = TRUE)}
}
\description{
This function takes in 2-digit level CPC codes and returns all corresponding 3-digit level CPC codes.
Input can be either numeric or character. Results are returned as characters
}
