% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cpc to hs.r
\name{gta_cpc_to_hs}
\alias{gta_cpc_to_hs}
\title{Convert CPC 2.1 codes into HS 2012 (6-digit level).}
\usage{
gta_cpc_to_hs(
     codes,
     cpc_digits = c(2, 3),
     as_list = FALSE,
     message = TRUE
)
}
\arguments{
\item{codes}{Supply the CPC codes you want to convert. Character and numeric is accepted.}

\item{as_list}{Indicate whether the output is a vector of uniquely converted codes (False) or as a list with each ... (copy from hs_vintage_converter())}

\item{message}{If TRUE, print conversion results, including the codes that could not be converted}

\item{cpc_digits}{If a vector of 2 digit cpc codes is supplied, specify \code{cpc_digits = 2}.}
}
\description{
This function converts vectors of CPC 2.1 codes (3 level) into a vector of 6-digit HS 2012 codes.
2 and 3 digit codes cannot be mixed as interpreting which level is ment can be ambiguous (eg. input of 17 could be "017" or "17" respectively)
}
