% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cumulative counts.R
\name{gta_cumulative_count}
\alias{gta_cumulative_count}
\title{GTA in force interventions counter}
\usage{
gta_cumulative_count(
     data,
     count_by = c("quarter", "year", "month")
)
}
\arguments{
\item{data}{requires a data frame obtained by running the \code{gta_data_slicer()}}

\item{count_by}{Specify whether to count by month, quarter or year.}
}
\value{
Outputs in force interventions by different given periods.
}
\description{
Computes number of interventions in force.
}
\examples{
# first, obtain a master dataset (eg. by running the data slicer)
master <- gta_data_slicer(affected.jurisdiction = "Switzerland", keep.affected = TRUE, keep.others = FALSE)
master |>
    gta_cumulative_count(count_by = "year")
}
