% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta cumulative counts.R
\name{gta_cumulative_counts}
\alias{gta_cumulative_counts}
\title{GTA in force interventions counter}
\usage{
gta_cumulative_counts(data, counts.by = "quarter")
}
\arguments{
\item{data}{requires a data frame obtained by running the \code{gta_data_slicer()}}

\item{counts.by}{Specify whether to count by month, quarter or year.'.}
}
\value{
Outputs in force interventions by different given periods.
}
\description{
Computes number of interventions in force.
}
\details{
all gta_data_slicer parameters are permissible.'.
}
