% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta intervention duration.r
\name{gta_intervention_duration}
\alias{gta_intervention_duration}
\title{GTA intra-year duration calculator}
\usage{
gta_intervention_duration(
  data = NULL,
  data.path = "data/master.rds",
  years = NULL,
  current_year_todate = TRUE
)
}
\arguments{
\item{data}{Allows to pass the data frame with interventions directly to the function. This is advisable if the function is run multiple times and avoids
a repeated loding of the data frame within the function.}

\item{years}{The calendar years for which to calculate the shares. Calculation includes interventions based on enforcement status, not implementation date i.e. if you start in 2010, this function will also work with interventions implemneted in 2009 but still in force in 2010. Specify as c(start.year, end.year). Default is c(2008,CURRENT.YEAR).}

\item{current_year_todate}{Should the duration for the current year, if included, be calculated as 'share of year to date' (TRUE) or 'share of entire current year' (FALSE). Default is TRUE.}

\item{data_path}{Specifies where the GTA data file is located (Default: 'data/master_plus.Rdata')}
}
\description{
Computes the share of each calendar year that each intervention was in force.
}
