% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta setwd.R
\name{gta_setwd}
\alias{gta_setwd}
\title{Sets your working directory to that used by the GTA data team.}
\usage{
gta_setwd(gdrive = NA)
}
\value{
A common working directory
}
\description{
Sets your working directory to that used by the GTA data team.
}
\examples{
# if your google drive .. on your computer starts with H:
gta_setwd(gdrive = "H")
# if your google drive .. on your computer starts with G:
gta_setwd(gdrive = "G")
# etc...
}
