% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta un code vector.R
\name{gta_un_code_vector}
\alias{gta_un_code_vector}
\title{Create a vector of UN country codes
Returns a vector of codes based on country or group names.}
\usage{
gta_un_code_vector(countries = "all")
}
\arguments{
\item{countries}{Vector with either country names, country group names or UN codes. Do not mix UN codes with country or group names. Please use function also for UN codes to ensure the ones you supply are consistent with the ones used by the GTA.}
}
\value{
A vector of UN country codes that is consistent with those used by the GTA.
}
\description{
Create a vector of UN country codes
Returns a vector of codes based on country or group names.
}
